move interfaces list and config into section
